added simple method to create a second window and place some ui elements inside. Basic usage example is shown in plugins/examples/holo_raycast.cxx Simply use cgv::gui::application to create a new window w of type "generic:align" or "generic:tab", use w->get_inner_group() to add elements like controls, rebinding their value_change signal to on_set and finally call w->show().
Also eliminated some warning.

+++ libs/cgv_app/transformation_gizmo.cxx
@@ -26,7 +26,7 @@ transformation_gizmo::transformation_gizmo() {
 	// calculate ring points for rotation handles
 	for(size_t i = 0; i < _ring_segment_count; ++i) {
 		float t = static_cast<float>(i) / static_cast<float>(_ring_segment_count - 1);
-		t *= 2.0f * M_PI;
+		t *= 2.0f * float(M_PI);
 		_ring_points.push_back({ std::cos(t), std::sin(t) });
 	}
 	_ring_points.back() = _ring_points.front();
@@ -234,7 +234,7 @@ void transformation_gizmo::create_geometry() {
 	switch(_interaction_feature) {
 	case InteractionFeature::kAxis:
 		if(use_axes) {
-			size_t base_idx = 2.0f * axis_idx;
+			size_t base_idx = 2 * axis_idx;
 			if(has_translation && has_scale && _interaction_mode == Mode::kScale || has_translation != has_scale) {
 				saturate_color(_cones.colors[base_idx]);
 				saturate_color(_cones.colors[base_idx + 1]);
@@ -272,8 +272,8 @@ void transformation_gizmo::create_geometry() {
 
 void transformation_gizmo::draw_geometry(context& ctx) {
 	// Pass the y view angle to the renderers so pixel measurements are computed correctly
-	_rectangle_renderer.set_y_view_angle(get_view()->get_y_view_angle());
-	_sphere_renderer.set_y_view_angle(get_view()->get_y_view_angle());
+	_rectangle_renderer.set_y_view_angle(float(get_view()->get_y_view_angle()));
+	_sphere_renderer.set_y_view_angle(float(get_view()->get_y_view_angle()));
 
 	// Since we use scaling in the modelview matrix we also need to adjust the pixel measures
 	// in the render styles based on the scale of the gizmo.
@@ -357,7 +357,7 @@ bool transformation_gizmo::intersect(const cgv::math::ray3& ray) {
 
 	if(_mode == Mode::kModel) {
 		cylinder_count += 3;
-		for(size_t i = 0; i < 3; ++i) {
+		for(int i = 0; i < 3; ++i) {
 			cylinders[i + 3].first[i] = axis_length + 2.0f * _handle_size;
 			cylinders[i + 3].second[i] = axis_length + 4.0f * _handle_size;
 		}
@@ -377,7 +377,7 @@ bool transformation_gizmo::intersect(const cgv::math::ray3& ray) {
 	// test rectangles for planes
 	if(_mode == Mode::kTranslation || _mode == Mode::kScale) {
 		float t = std::numeric_limits<float>::max();
-		for(size_t i = 0; i < 3; ++i) {
+		for(int i = 0; i < 3; ++i) {
 			vec3 position = { 0.5f };
 			position[i] = 0.0f;
 			if(cgv::math::ray_axis_aligned_rectangle_intersection(ray, position, { _plane_size }, static_cast<int>(i), t))
@@ -554,7 +554,7 @@ bool transformation_gizmo::drag(const cgv::math::ray3& ray) {
 
 			// check for side and bring angle in range [0,2pi];
 			if(dot(plane_tangent, end_dir) < 0.0f)
-				angle = 2.0f * M_PI - angle;
+				angle = 2.0f * float(M_PI) - angle;
 
 			vec3 rotaton_axis = _interaction_feature == InteractionFeature::kCenter ? _interaction_plane.normal : axis;
